# static/openapi.yaml
openapi: 3.1.0
info:
  title: IoT and IM Microservices API (Synchronous)
  description: |
    Formal API specification for the Synchronous Microservices Architecture involving:
    - **IM Services**: Telegram (/IMTelegram) and Line (/IMLine)
    - **IoT Broker**: Central logic and message routing (/IoTBroker)
    - **IoT Device Controller**: Proxy for physical devices (/IoTDevice)
    
    This API facilitates synchronous HTTP communication between chat platforms and IoT devices.
  version: 2.1.0
  contact:
    name: Rocer52
    email: batte9306@gmail.com
servers:
  - url: http://localhost:5000
    description: IM Telegram Service
  - url: http://localhost:5001
    description: IM Line Service
  - url: http://localhost:5002
    description: IoT Broker Service
  - url: http://localhost:5003
    description: IoT Device Controller Service
  - url: http://localhost:5010
    description: ESP32 Virtual Device
  - url: http://localhost:5011
    description: Raspberry Pi Virtual Device
...
paths:
...
  /IoTDevice/Enable:
    get:
      tags:
        - IoTDevice
      summary: Enable Device
      description: Send an enable command to the target physical/virtual device.
      operationId: enableDevice
      parameters:
        - $ref: '#/components/parameters/device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/platform'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Device enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '500':
          description: Device unreachable or error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /IoTDevice/Disable:
    get:
      tags:
        - IoTDevice
      summary: Disable Device
      description: Send a disable command to the target physical/virtual device.
      operationId: disableDevice
      parameters:
        - $ref: '#/components/parameters/device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/platform'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Device disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '500':
          description: Device unreachable or error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /IoTDevice/GetStatus:
    get:
      tags:
        - IoTDevice
      summary: Get Device Status
      description: Query the current status of the target physical/virtual device.
      operationId: getDeviceStatus
      parameters:
        - $ref: '#/components/parameters/device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/platform'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '500':
          description: Device unreachable or error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Invalid request or internal error"
      required:
        - message
...
    DeviceResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Device enabled"
        state:
          type: string
          example: "on"
        device_id:
          type: string
          example: "esp32_light_001"
      required:
        - status
        - message
        - state
        - device_id
...
  parameters:
    device_id:
      name: device_id
      in: query
      description: Target Device ID
      required: true
      schema:
        type: string
        enum:
          - esp32_light_001
          - esp32_fan_002
          - raspberrypi_light_001
          - raspberrypi_fan_002
        example: "esp32_light_001"
    chat_id:
      name: chat_id
      in: query
      description: Chat ID of the requester
      required: false
      schema:
        type: string
        example: "123456789"
    platform:
      name: platform
      in: query
      description: Originating platform
      required: false
      schema:
        type: string
        enum: [telegram, line]
        default: telegram
        example: "telegram"
    user_id:
      name: user_id
      in: query
      description: User ID
      required: false
      schema:
        type: string
        example: "123456789"
    username:
      name: username
      in: query
      description: User display name
      required: false
      schema:
        type: string
        example: "user123"
    bot_token:
      name: bot_token
      in: query
      description: Bot token (internal use)
      required: false
      schema:
        type: string
        example: "123456789:ABCDEF_ExampleBotToken"